requirements.txt
google-generativeai
requests
mcp
google-genai
Pathlib